requirements.txt
FastAPI
Uvicorn[standard]
SQLAlchemy
python-dotenv
psycopg2-binary
pydantic
pytest
testcontainers[postgresql]